refactor backend ai tools modules
