+++ apps/backend/src/aiTools/sqlDialect/types.ts
