+++ apps/backend/src/aiTools/toolContract/sqlToolContract.ts
