+++ apps/backend/src/aiTools/toolContract/sqlToolLimits.ts
@@ -2,7 +2,7 @@
  * Canonical SQL tool limits for the TypeScript runtimes.
  *
  * Keep these values aligned with:
- * - `apps/backend/src/aiTools/agentSqlShared.ts`
+ * - `apps/backend/src/aiTools/agentSql/shared.ts`
  * - `apps/ios/Flashcards/Flashcards/AI/AIChatTypes.swift`
  */
 export const MAX_SQL_RECORD_LIMIT = 100;
